requirements.txt
python-telegram-bot==13.15
google-generativeai==0.3.2